SSUT-344: Fix NationalityOfTransport nav
It should always redirect to the identity page for the selected mode of
transport; previously it was always going to unaccompanied roro.

While I'm here, fix a mistaken introduced to TransportModePage's nav
which weakened our guarantees for handling each type of transport mode
and introduced an inexhaustive match, instead explicitly sending the
user to journey recovery if the match fails.