SSUT-287/SSUT-283: Party models + serialisers
All fields referring to a person/company are identified by either eori
or name + address, so we can use a single Party model in the XML to
refer to each of these parties.

These fields are serialised with different field names, though, so we
need to provide the correct set of field names for each party
serialised. Create a generic Party serialisation format which takes a
set of fields in order to achieve this in a DRY way.

Also includes a hotfix for PackagesFormats: standardKindsOfPackages was
used in the decoder, rather than allKindsOfPackage. This failed the
tests when a non-standard package type was picked.