SSUT-422: Fix uses of Gen.XXXStr + take(n) (#101)
Gen.alphaStr / Gen.alphaNumStr etc. have the possibility of producing
very short strings or even zero-length strings, which make this approach
of creating strings of certain length, or even of certain max length,
unreliable, and thios caused a controller test to fail rarely due to
zero-length strings being submitted to the form when valid data was
expected.

Fix all uses of this in ModelGenerators to the more reliable Gen.listOfN
+ Gen.XXXChar approach.
Fix uses of this in a handful of tests to the same, or else reusing the
utils in Generators (stringsOfMaxLength etc.) where applicable.
Also replaced a couple of generated strings with arbitrary[GbEori] since
they were generating a GB EORI anyway.

While I'm here I updated the definition of the arbitrary MessageSender,
since this was one of the affected arbitrary generators, and updated the
scaladoc on this model to reflect what "branch" really is, since we've
since clarified the meaning of this field.