SSUT-470: Add a declaration client
Add a basic DeclarationConnector (API client), currently capable of
handling submissions and returning the CorrelationId from a successful
response. Bad responses will fail the future with  meaningful
exceptions.

Trait can be extended for unit testing, concrete client is an
injectable type, taking FrontendAppConfig and deriving its config from
there.

Add decoding for CorrelationId, extracting it from successful submission
/ amendment responses

Update Service to allow providing a url path prefix as well, as some
APIs may have a static prefix we want to always provide. Also swap
baseUrl from a string to a URL, since the http client recommends we
always use URL to ensure safe URL construction. Use the provided
interpolator to do this.

Make sure Integration extends Test, so that we can still use Test utils
from within the it tests.

Add a test dependency on the http-verbs testkit for ease of testing the
client with wiremock.

Add XmlPayloadFixtures to easily handle sample XML payloads we need for
testing, and add an example response payload for a submission, sourced
from the S&S API docs

Add the client to the CheckYourAnswersController, and adjust the route,
controller, and view to allow submitting the page here and ultimately
we'll then trigger declaration submission. For the moment, just return OK