SSUT-524: Consignor may be in top-level payload
If the consignor is the same for all goods items, we're obliged to
provide it at the top level fo the payload, and must not provide it in
the individual good items.

To facilitate this, add consignor as an optional field to the root
model, and change consignor to optional on the goods item model.