SSUT-530: Allow storing submission events/outcomes
Add models representing declaration submissions and amendments, along
with the corresponding outcomes

Add a list of such events to UserAnswers, defaulting to an empty list
until they happen

Add a method to SessionRepository to allow storing a declaration
event.

Storing outcomes is going to be slightly more complex as we'll need to
do it by correlation ID, and therefore need to be able to find a
document by one of its correlation IDs. This work will be in the
submissions microservice which is being built already, so I won't
duplicate effort here.