Fix automerge bug around Predec
Nick recently changed the fields in the Predec model, and then
Salman/James added a model generator for this model based on the old
version.

The result is git was happy to automerge the changes, which occurred in
different places, but that automerge broke compilation.

This commit properly resolves the conflict.