SSUT-400: Adjust transport identity fields
The fields "transport identity" and "conveyance reference number" are
both optional, as one or the other may be omitted depending on the mode
of transport, following some fairly complex business rules.

Conveyance reference number refers to air or maritime transport means,
so it more properly belongs in the TransportDetails section

Make these model changes and update the associated formats to continue
to format them the same way.

This is part 1 in the changes required to satisfy SSUT-400