SSUT-566: Limit add-to-list sections (#139)
The final payload restricts how many goods items, documents, containers, seals etc. can be present. This PR adds limits to various add-to-list sections in the journey. The limitations work in two ways:

On the various "add another" pages, the "do you want to add another" question is replaced with guidance that the max number have already been added, and the controller, forces the answer to false.
On the various question pages inside a section, a filter is applied to the onPageLoad and onSubmit methods to redirect to Journey Recovery if the index is too high.
This logic has been applied to all of the sections that result in data being added to the payload. It has not been applied to consignors, consignees, notified parties, places of loading or places of unloading - that could be done in a separate PR, but would not cause any issues if they added too many (since they would be prevented from using some of them by the fact the number of goods items they can add is restricted).

Overall documents and item-level documents should ideally interact; the overall document limit should be reduced by the highest number of documents added to goods items, and the goods item limit reduced by the number of overall documents. This has not been implemented in this PR, and each list is separately limited to 99. That logic should be introduced separately, although it's an edge case that users would need or want to add that many doucments.